+++ docs/reference/channels/backends/base.rst
@@ -1,0 +1,5 @@
+base
+=====
+
+.. automodule:: litestar.channels.backends.base
+    :members:
