+++ docs/reference/channels/backends/index.rst
@@ -1,0 +1,8 @@
+backends
+========
+
+.. toctree::
+
+    base
+    memory
+    redis
