+++ docs/reference/channels/backends/memory.rst
@@ -1,0 +1,5 @@
+memory
+======
+
+.. automodule:: litestar.channels.backends.memory
+    :members:
