+++ docs/reference/channels/backends/redis.rst
@@ -1,0 +1,5 @@
+redis
+=====
+
+.. automodule:: litestar.channels.backends.redis
+    :members:
