--- docs/reference/channels/base.rst
@@ -1,5 +1,0 @@
-base
-=====
-
-.. automodule:: litestar.channels.base
-    :members:
