+++ docs/reference/channels/index.rst
@@ -3,7 +3,5 @@ channels
 
 .. toctree::
 
-    base
     plugin
-    memory
-    redis
+    backends/index
