--- docs/reference/channels/memory.rst
@@ -1,5 +1,0 @@
-memory
-======
-
-.. automodule:: litestar.channels.memory
-    :members:
