--- docs/reference/channels/redis.rst
@@ -1,5 +1,0 @@
-redis
-=====
-
-.. automodule:: litestar.channels.redis
-    :members:
