+++ litestar/channels/__init__.py
@@ -1,4 +1,4 @@
-from .base import ChannelsBackend
+from .backends.base import ChannelsBackend
 from .plugin import ChannelsPlugin, Subscriber
 
 __all__ = ("ChannelsPlugin", "ChannelsBackend", "Subscriber")
